requirements.txt
torch>=1.8.0
torchvision>=0.9.0
matplotlib>=3.3.0
tqdm>=4.48.0
Pillow>=8.0.0
huggingface-hub>=0.0.19
kagglehub>=0.1.0 